Refactored e2e command to be more in line with manager
